requirements.txt
alembic==1.17.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
